chore(tui): extract the wordmark, fix status-bar clipping, update docs
`Wordmark` comes out of WelcomeScreen so both entry screens draw the same
banner and its width fallback lives in one place.

The status bar clipped its breadcrumb to "Goa" between 100 and 118
columns; the breadcrumb no longer shrinks, so the hints truncate instead.

Docs updated across README, CLAUDE.md, the CLI skill and the landing
page's usage guide for goal mode, the stream chords, `/export` and the
config behaviour.
