+++ packages/cli/src/ui/ink/components/TokenSummary.tsx
@@ -1,12 +1,16 @@
 import React from 'react';
-import { Box, Text } from 'ink';
+import { Box, Text, useStdout } from 'ink';
 import { COLORS } from '../theme.js';
 import type { TokenSummaryData } from '../store.js';
+import { useRunWidth } from '../width.js';
 
-const W = 64;
+const MAX_W = 64;
 
 /** Final token / cost / model panel, committed after the result box. */
 export function TokenSummary({ data }: { data: TokenSummaryData }) {
+  const { stdout } = useStdout();
+  // marginLeft(2) plus a cell of slack, so a narrow pane clips nothing.
+  const W = Math.min(MAX_W, Math.max(20, useRunWidth(stdout?.columns) - 3));
   const total = data.input + data.output;
   if (total === 0) return null;
